updates to PaymentEntity.kt

+++ app/src/main/java/com/example/invoicemaker/data/local/entity/PaymentEntity.kt
@@ -1,10 +1,10 @@
+// data.local.entity.PaymentEntity.kt
 package com.example.invoicemaker.data.local.entity
 
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import java.math.BigDecimal
 
 @Entity(
     tableName = "payments",
@@ -21,7 +21,7 @@ import java.math.BigDecimal
 data class PaymentEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val invoiceId: Long,
-    val amount: BigDecimal,
+    val amount: Double,
     val date: Long,
     val method: String? = null,
     val note: String? = null
